collector: move filters? to a django command
