--- collectr/collector/filters.py
@@ -1,25 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-    Collector filtering url system.
-
-"""
-import sys
-import pprint
-
-class Filter(object):
-    """
-
-    """
-
-    def __init__(self:
-        pass
-
-
-def test():
-    """some quick tests"""
-    pass
-
-if __name__ == '__main__':
-    test()
-
-
